sixth attempt

+++ src/main/java/com/driver/StudentRepository.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class StudentRepository {
     HashMap<String,Student> studentDb= new HashMap<>();
     HashMap<String,Teacher> teacherDb=new HashMap<>();
-    HashMap<String,HashSet<String>>teacherStudentDb=new HashMap<>();
+    HashMap<String,List<String>>teacherStudentDb=new HashMap<>();
 
     public void addStudent(Student student){
         String key=student.getName();
@@ -26,31 +26,38 @@ public class StudentRepository {
     }
 
     public void addStudentTeacherPair(String studentName, String teacherName){
-        HashSet<String >students=teacherStudentDb.get(teacherName);
+
+        if(teacherStudentDb.containsKey(teacherName)) {
+
+            List<String>students = teacherStudentDb.get(teacherName);
+            students.add(studentName);
+            teacherStudentDb.put(teacherName,students);
+            return;
+        }
+        List<String> students =new ArrayList<>();
         students.add(studentName);
         teacherStudentDb.put(teacherName,students);
     }
 
     public Student getStudentByName(String name){
-        Student student=studentDb.get(name);
-        return student;
+        if(studentDb.containsKey(name)) {
+            Student student = studentDb.get(name);
+            return student;
+        }
+        return null;
     }
     public Teacher getTeacherByName(String name){
-        Teacher teacher=teacherDb.get(name);
-        return teacher;
+        if(teacherDb.containsKey(name)) {
+            Teacher teacher = teacherDb.get(name);
+            return teacher;
+        }
+        return null;
     }
     public List<String> getStudentsByTeacherName(String teacherName){
-
-        List<String>studentsist=new ArrayList<>();
-        if(!(teacherStudentDb.containsKey(teacherName)))
-            return studentsist;
-        HashSet<String>students=teacherStudentDb.get(teacherName);
-
-
-        for(String student:students){
-            studentsist.add(student);
+        if(teacherStudentDb.containsKey(teacherName)){
+            return teacherStudentDb.get(teacherName);
         }
-        return studentsist;
+        return new ArrayList<>();
     }
     public List<String> getAllStudents(){
         List<String>studentsList=new ArrayList<>();
@@ -59,26 +66,33 @@ public class StudentRepository {
         }
         return studentsList;
     }
-    public void deleteTeacherByName(String teacherName){
-
-
-
-            for(String student:teacherStudentDb.get(teacherName)){
-                studentDb.remove(student);
+    public void deleteTeacherByName(String teacherName) {
+        if (teacherDb.containsKey(teacherName)) {
+            teacherDb.remove(teacherName);
+        }
+        if (teacherStudentDb.containsKey(teacherName)) {
+            for (String student : teacherStudentDb.get(teacherName)) {
+                if(studentDb.containsKey(student)) {
+                    studentDb.remove(student);
+                }
             }
-        teacherStudentDb.remove(teacherName);
-        teacherDb.remove(teacherName);
-
-
+            teacherStudentDb.remove(teacherName);
+        }
     }
     public void deleteAllTeachers(){
-       for(String teacherName:teacherDb.keySet()){
-           for(String student:teacherStudentDb.get(teacherName)){
-               studentDb.remove(student);
-           }
-           teacherStudentDb.remove(teacherName);
-           teacherDb.remove(teacherName);
+
+        for(String teacherName:teacherDb.keySet()){
+            teacherDb.remove(teacherName);
+            if (teacherStudentDb.containsKey(teacherName)) {
+                for (String student : teacherStudentDb.get(teacherName)) {
+                    if(studentDb.containsKey(student)) {
+                        studentDb.remove(student);
+                    }
+                }
+                teacherStudentDb.remove(teacherName);
+            }
        }
+        teacherStudentDb.clear();
